move more functions to release.py
