+++ fab/fabfile.py
@@ -379,30 +379,6 @@ def mail_admins(subject, message):
         })
 
 
-@roles(ROLES_DB_ONLY)
-def record_successful_deploy():
-    with cd(env.code_current):
-        env.deploy_metadata.tag_commit()
-        sudo((
-            '%(virtualenv_current)s/bin/python manage.py '
-            'record_deploy_success --user "%(user)s" --environment '
-            '"%(environment)s" --url %(url)s --mail_admins'
-        ) % {
-            'virtualenv_current': env.virtualenv_current,
-            'user': env.captain_user or env.user,
-            'environment': env.environment,
-            'url': env.deploy_metadata.diff_url,
-        })
-
-
-
-@roles(ROLES_ALL_SRC)
-@parallel
-def record_successful_release():
-    with cd(env.root):
-        files.append(RELEASE_RECORD, str(env.code_root), use_sudo=True)
-
-
 @task
 def hotfix_deploy():
     """
@@ -427,7 +403,7 @@ def hotfix_deploy():
         raise
     else:
         silent_services_restart(use_current_release=True)
-        execute(record_successful_deploy)
+        execute(release.record_successful_deploy)
 
 
 def _confirm_translated():
@@ -442,9 +418,9 @@ def _confirm_translated():
 @task
 def setup_release():
     deploy_ref = env.deploy_metadata.deploy_ref  # Make sure we have a valid commit
-    execute_with_timing(create_code_dir)
+    execute_with_timing(release.create_code_dir)
     execute_with_timing(release.update_code, deploy_ref)
-    execute_with_timing(update_virtualenv)
+    execute_with_timing(release.update_virtualenv)
 
     execute_with_timing(copy_release_files)
 
@@ -483,7 +459,7 @@ def _deploy_without_asking():
         # hard update of manifest.json since we're about to force restart
         # all services
         execute_with_timing(staticfiles.update_manifest)
-        execute_with_timing(clean_releases)
+        execute_with_timing(release.clean_releases)
     except PreindexNotFinished:
         mail_admins(
             " You can't deploy yet",
@@ -501,24 +477,15 @@ def _deploy_without_asking():
         silent_services_restart()
         raise
     else:
-        execute_with_timing(update_current)
+        execute_with_timing(release.update_current)
         silent_services_restart()
-        execute_with_timing(record_successful_release)
-        execute_with_timing(record_successful_deploy)
+        execute_with_timing(release.record_successful_release)
+        execute_with_timing(release.record_successful_deploy)
 
 
 @task
-@roles(ROLES_ALL_SRC)
-@parallel
 def update_current(release=None):
-    """
-    Updates the current release to the one specified or to the code_root
-    """
-    if ((not release and not files.exists(env.code_root)) or
-            (release and not files.exists(release))):
-        utils.abort('About to update current to non-existant release')
-
-    sudo('ln -nfs {} {}'.format(release or env.code_root, env.code_current))
+    execute(release.update_current, release)
 
 
 @task
@@ -535,13 +502,6 @@ def unlink_current():
 
     if files.exists(env.code_current):
         sudo('unlink {}'.format(env.code_current))
-
-
-@task
-@roles(ROLES_ALL_SRC)
-@parallel
-def create_code_dir():
-    sudo('mkdir -p {}'.format(env.code_root))
 
 
 @parallel
@@ -639,9 +599,9 @@ def rollback():
 
     if all(exists.values()):
         print blue('Updating current and restarting services')
-        execute(update_current, unique_release)
+        execute(release.update_current, unique_release)
         silent_services_restart(use_current_release=True)
-        execute(mark_last_release_unsuccessful)
+        execute(release.mark_last_release_unsuccessful)
     else:
         print red('Aborting because not all hosts have release')
         exit()
@@ -652,14 +612,6 @@ def rollback():
 def get_number_of_releases():
     with cd(env.root):
         return int(sudo("wc -l {} | awk '{{ print $1 }}'".format(RELEASE_RECORD)))
-
-
-@roles(ROLES_ALL_SRC)
-@parallel
-def mark_last_release_unsuccessful():
-    # Removes last line from RELEASE_RECORD file
-    with cd(env.root):
-        sudo("sed -i '$d' {}".format(RELEASE_RECORD))
 
 
 @roles(ROLES_ALL_SRC)
@@ -677,39 +629,11 @@ def get_previous_release():
 
 
 @task
-@roles(ROLES_ALL_SRC)
-@parallel
 def clean_releases(keep=3):
     """
     Cleans old and failed deploys from the ~/www/<environment>/releases/ directory
     """
-    releases = sudo('ls {}'.format(env.releases)).split()
-    current_release = os.path.basename(sudo('readlink {}'.format(env.code_current)))
-
-    to_remove = []
-    valid_releases = 0
-    with cd(env.root):
-        for index, release in enumerate(reversed(releases)):
-            if (release == current_release or release == os.path.basename(env.code_root)):
-                valid_releases += 1
-            elif (files.contains(RELEASE_RECORD, release)):
-                valid_releases += 1
-                if valid_releases > keep:
-                    to_remove.append(release)
-            else:
-                # cleans all releases that were not successful deploys
-                to_remove.append(release)
-
-    if len(to_remove) == len(releases):
-        print red('Aborting clean_releases, about to remove every release')
-        return
-
-    if os.path.basename(env.code_root) in to_remove:
-        print red('Aborting clean_releases, about to remove current release')
-        return
-
-    for release in to_remove:
-        sudo('rm -rf {}/{}'.format(env.releases, release))
+    execute(release.clean_releases)
 
 
 @task
@@ -772,39 +696,6 @@ def awesome_deploy(confirm="yes"):
         print('┻┻┻┻┻┻')
 
     _deploy_without_asking()
-
-
-@roles(ROLES_ALL_SRC)
-@parallel
-def update_virtualenv():
-    """
-    update external dependencies on remote host
-
-    assumes you've done a code update
-
-    """
-    _require_target()
-    requirements = posixpath.join(env.code_root, 'requirements')
-
-    # Optimization if we have current setup (i.e. not the first deploy)
-    if files.exists(env.virtualenv_current):
-        print 'Cloning virtual env'
-        # There's a bug in virtualenv-clone that doesn't allow us to clone envs from symlinks
-        current_virtualenv = sudo('readlink -f {}'.format(env.virtualenv_current))
-        sudo("virtualenv-clone {} {}".format(current_virtualenv, env.virtualenv_root))
-
-    with cd(env.code_root):
-        cmd_prefix = 'export HOME=/home/%s && source %s/bin/activate && ' % (
-            env.sudo_user, env.virtualenv_root)
-        # uninstall requirements in uninstall-requirements.txt
-        # but only the ones that are actually installed (checks pip freeze)
-        sudo("%s bash scripts/uninstall-requirements.sh" % cmd_prefix,
-             user=env.sudo_user)
-        sudo('%s pip install --timeout 60 --quiet --requirement %s --requirement %s' % (
-            cmd_prefix,
-            posixpath.join(requirements, 'prod-requirements.txt'),
-            posixpath.join(requirements, 'requirements.txt'),
-        ))
 
 
 @task
